+++ powerdns/provider.go
@@ -81,7 +81,6 @@ func Provider() *schema.Provider {
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
-			"powerdns_zone":                  resourcePDNSZone(),
 			"powerdns_view_zone_association": resourcePDNSViewZoneAssociation(),
 			"powerdns_network":               resourcePDNSNetwork(),
 			"powerdns_zone_metadata":         resourcePDNSZoneMetadata(),